refactor: querySelector로 교체하기
