+++ src/utils/selector.js
@@ -1,0 +1,2 @@
+const $ = (selector) => document.querySelector(selector);
+export default $;
